Create file for BankAccount interfaces and class
