+++ packages/webcomponents/src/api/BankAccount.ts
@@ -1,0 +1,12 @@
+
+export interface BankAccount {
+  id: string;
+  full_name: string;
+  bank_name: string;
+  account_number_last4: string;
+  routing_number: string;
+  country: string;
+  currency: string;
+  nickname: string;
+  account_type: string;
+}
