+++ packages/webcomponents/src/api/shared.ts
@@ -1,15 +1,3 @@
-export interface BankAccount {
-  id: string;
-  full_name: string;
-  bank_name: string;
-  account_number_last4: string;
-  routing_number: string;
-  country: string;
-  currency: string;
-  nickname: string;
-  account_type: string;
-}
-
 // Errors on [capital]/config/locales/en.yml
 // Add more errors as needed
 export const API_ERRORS = {
